Fix percentage display.
The percentage used display is not always correct when reviewing the
OLED display and output of df.  Fix the code so we get the calculated
percentage from df, which seems to make everything align.

Added argon-status.py so we can report on the status of the system
when we are not able to see the system itself.