fix get_fanspeed function.
The get_fanspeed function scans a list of temperatures, and
looks for the best match for the current temperature passed in and
the corresponding fanspeed.

The code had an issue where if the fanspeed was less than 25, it
returned 25% all the time.  So if you had a minimum temp value of say
35 degrees and a fan speed of 0, the function would return a fanspeed of
25% all the time.  This has been eliminated.

Also the temp_check function has been updatd to read in the
configuration file EVERY SINGLE TIME so that we can now tweek the fan
speeds on the fly.  Once a change is made, approximately 30 seconds
later the corresponding change will be made to the fan.  Note in
situations where the new fan speed is lower, it will take an additional
30 seconds to respond as the code is attemping to remove fluxuations in
fan speed.