get_fanspeed was not returning the proper value
get_fanspeed was returning the temperature of the device as a fanspeed,
this is why we were getting floating point values were integers were
supposed to be delivered.  Fixed the issue.

Also added the ability to get the temperature of my NVME device, not
sure if it will work for others.